Move ActTargetInfo.cs into separate document
